% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QR.R
\name{QR}
\alias{QR}
\title{QR method}
\usage{
QR(X)
}
\arguments{
\item{X}{is a matrix consisting of which variables should regressed on}
}
\value{
The regression variables coefficients are returned
}
\description{
QR method
}
\examples{

}
